refactor(#189): ui.ts -> option.ts로 네이밍 변경
