+++ src/components/NavBar/index.tsx
@@ -13,7 +13,7 @@ import { UserData } from '@customTypes/user';
 
 import { ROUTES } from '@constants/path';
 import { LOGO_IMAGE } from '@constants/common/image';
-import { NAVBAR_DROPDOWN_LABELS } from '@constants/common/ui';
+import { NAVBAR_DROPDOWN_LABELS } from '@constants/common/option';
 import { ACCESS_TOKEN_KEY } from '@constants/api';
 
 import { useModalStore } from '@stores/useModalStore';
