+++ src/pages/MainLayout/Footer/index.tsx
@@ -1,6 +1,6 @@
 import FooterLabel from '@pages/MainLayout/Footer/FooterLabel';
 import { LOGO_IMAGE } from '@constants/common/image';
-import { MENU_LIST } from '@constants/common/ui';
+import { MENU_LIST } from '@constants/common/option';
 import S from './style';
 
 export default function Footer() {
