+++ src/pages/Study/StudyList/index.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import useGetStudyList from '@hooks/study/useGetStudyList';
 import { BANNER_IMAGES } from '@constants/common/image';
-import { ACCESS_STATUS_MAP_ID, STUDY_PAGE_SIZE } from '@constants/common/ui';
+import { ACCESS_STATUS_MAP_ID, STUDY_PAGE_SIZE } from '@constants/common/option';
 import { StudyData } from '@customTypes/study';
 import PageButton from 'src/components/_common/molecules/PageButton';
 import StudyCard from 'src/components/Study/StudyCard';
